pyproject.toml
[project]
name = "exchange-rate-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "fastapi[standard]>=0.121.1",
    "google-auth>=2.43.0",
    "google-cloud-bigquery>=3.38.0",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.0",
    "ruff>=0.14.4",
]
